Refactor: move component realted to buttons
