+++ src/components/Header/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { AppstoreOutlined, BellOutlined, QuestionCircleOutlined, DownOutlined, UserOutlined } from "@ant-design/icons";
-import Button from "@components/Button";
+import Button from "@components/buttons/Button";
 import * as S from "./style";
 
 function Header() {
