+++ src/components/Modals/Modal/index.tsx
@@ -1,12 +1,12 @@
 import React, { Suspense, useCallback } from "react";
-import { CloseModalState, ModalState } from "@/interfaces/modal";
+import { type CloseModalState, type ModalState } from "@/interfaces/modal";
 import * as S from "./style";
 
 interface Props<P> extends ModalState<P> {
   onClose: (state: CloseModalState) => void;
 }
 
-const Modal = <P extends {}>({ component: Component, index, onClose, props, options }: Props<P>) => {
+const Modal = <P extends Record<string,unknown>>({ component: Component, index, onClose, props, options }: Props<P>) => {
   const handleClose = useCallback(() => {
     onClose({ component: Component, index });
   }, [Component, index, props]);
